… Cargo.lock
… Cargo.toml
… crates/gitbutler-branch-actions/Cargo.toml
… crates/gitbutler-cli/src/command/mod.rs
… crates/gitbutler-cli/src/command/vbranch.rs
… crates/gitbutler-cli/src/main.rs
… crates/gitbutler-command-context/src/lib.rs
… crates/gitbutler-oplog/Cargo.toml
… crates/gitbutler-oplog/src/oplog.rs
… crates/gitbutler-oxidize/src/lib.rs
… crates/gitbutler-repo/Cargo.toml
… crates/gitbutler-repo/src/repository_ext.rs
… crates/gitbutler-repo/tests/create_wd_tree.rs
… crates/gitbutler-testsupport/src/testing_repository.rs