The `gitoxide` merge-tree implementation is faster and better in quality, so it should be used where applicable. This PR attempts to find these places and make the switch.

The goal is also to provide performance metrics, even though it's not expected we will see great numbers here as merge-tree is a smaller portion of the total work to be done.

Follow-up on #5411 .

### Tasks

Use `merge-tree` in…

* [x] Apply and Unapply branch
* [x] unapply ownership
* [x] back to integration
* [x] Oplog Listing (3x faster, 3,5x for merge-trees)
* [x] checkout branch trees
* [x] upstream integration statuses
* [x] use latest `gix` for nicer API
* [x] be sure to switch `gix` to the one from `main`

### Notes for the Reviewer

* All uses that deal with `resolve_index()` have been postponed for now, as it's probably easiest to implement this directly in the `gitoxide` merge tree function, for achieving the best possible quality. In theory, it might already work doing a post-process on the conflict entries, but that would be harder to get right.
* There are a couple of other uses in `gitbutler-cherry-pick` and `gitbutler-edit-mode` that rely on the index quite a bit, and even checkout an index. But it *seems* that [it doesn't rely on the index to show merge-conflicts](https://github.com/gitbutlerapp/gitbutler/blob/068833059d6dbaad32b6aebe4d6863ad55333a38/crates/gitbutler-edit-mode/src/lib.rs#L102).

### Follow-Up

* Improve `gix` merge trees APIs, see https://github.com/GitoxideLabs/gitoxide/pull/1651
* Use `gix-merge` in all the remaining places for performance and merge-quality

### Apply and Unapply

#### Before

When *applying*, merge-trees takes a whopping **72ms**. It's clearly not the bottleneck, with a lot of time being spent in checkout.

<img width="944" alt="Screenshot 2024-11-03 at 14 22 36" src="https://github.com/user-attachments/assets/57399569-333c-480b-846e-6102c97da2cf">

When *unapplying*, it now clocks in at **139ms**, so once again isn't going to change anything when replaced. Checking out files takes a long time, comparatively.

<img width="950" alt="Screenshot 2024-11-03 at 14 24 46" src="https://github.com/user-attachments/assets/5c85dfd9-5427-4654-a3ba-71c9c6251ec8">


#### After

Like expected, there isn't much of a relevant change, with both merges during unapply and apply clocking in to ~90ms. Yes, it *seems*  a bit faster, but is still far from flexing any muscle.

<img width="920" alt="Screenshot 2024-11-03 at 16 05 34" src="https://github.com/user-attachments/assets/e0c74971-37da-4c85-b570-81aa9dcf3e3a">

### Unapply Ownership

#### Before 

A tiny portion of 66ms of the 4s runtime is related to merging trees. Since it's the same repository  (GitLab) the after should look very similar.

<img width="953" alt="Screenshot 2024-11-03 at 17 38 25" src="https://github.com/user-attachments/assets/84f77d17-0336-44fd-8606-222c9bb44777">


#### After

It's basically the same at 64ms.

<img width="940" alt="Screenshot 2024-11-03 at 17 38 59" src="https://github.com/user-attachments/assets/c81f2c71-7318-43a6-b11b-365804b1d19b">''

### Set Base Branch / Back to Integration

#### Before

**112ms**

<img width="935" alt="Screenshot 2024-11-03 at 19 30 17" src="https://github.com/user-attachments/assets/98f56c64-afcb-44a9-80ce-1d4b3b00eb0b">

#### After

**93ms**

<img width="938" alt="Screenshot 2024-11-03 at 19 34 21" src="https://github.com/user-attachments/assets/9b5d0348-7239-41b4-8e0b-0bcb8f8b3815">

### Oplog Listing

#### Before

Listing takes 22s, and one full diff takes 7s, which is duplicated for an additional stats. Merging of trees takes **1.93s**.

<img width="946" alt="Screenshot 2024-11-03 at 19 52 45" src="https://github.com/user-attachments/assets/f94bef9b-63d9-45e4-8746-cdbe464ac8d2">

#### After replacing merge-trees

Merging trees now take **585ms**,  3.3x of what was before.

<img width="939" alt="Screenshot 2024-11-03 at 20 28 31" src="https://github.com/user-attachments/assets/63d59962-3f73-48c1-9790-2ac68a3f5a78">

#### After using `gitoxide` for the diffing

The whole operation is down to 7.3s, a 3x improvement! This is also due to not duplicating work.

<img width="946" alt="Screenshot 2024-11-04 at 08 14 31" src="https://github.com/user-attachments/assets/d1c6f3d0-fbfd-4a0c-86b6-630d827b98b5">


It's noticeable how the TOML deserialisation takes 2s, which is a lot in comparison to the total runtime. But like I mentioned on Discord, I think there is ways to speed this up by not precomputing everything, instead of parallelising.

### Checkout Branch Trees

#### Before

Merging in total takes about **300ms** with two trivial branches. Like noticed before, writing the index to a tree is very expensive.

<img width="958" alt="Screenshot 2024-11-04 at 11 04 06" src="https://github.com/user-attachments/assets/16eb6cd4-3ad6-4cc5-8302-62516b43cdd7">

#### After

Now the same task takes **140ms** for merging and writing the tree. Notably, writing the tree now takes 6ms , 29x faster than having to write a whole index.

<img width="956" alt="Screenshot 2024-11-04 at 12 13 51" src="https://github.com/user-attachments/assets/b7977933-9d60-449e-9197-73a2419ad1cd">


### Upstream Integration Statuses

#### Before

As often, most time is spent elsewhere (and duplicated in `stack_series()`), but that aside we are looking at **1.09s** for merging trees in that function.

<img width="935" alt="Screenshot 2024-11-04 at 16 11 10" src="https://github.com/user-attachments/assets/850dc143-4161-4f43-ac26-314f352b6434">

#### After

Now it clocks in at 640ms, at 1.7x.

<img width="969" alt="Screenshot 2024-11-04 at 19 21 50" src="https://github.com/user-attachments/assets/7ad0c403-e9e4-4dfb-ae79-5ba729be875d">